ObjHits_GetPriorityHit(WithPosition): (s8)bestPriority for signed extsb compare (#15) (2 fns ->100)
The priority compare (s8)hitPriority < bestPriority sign-extended only the
left operand (extsb) and zero-extended the u8 bestPriority (clrlwi); target
sign-extends both. Cast bestPriority to s8 at the compare (storage stays u8).
ObjHits_GetPriorityHit 98.8->100, ObjHits_GetPriorityHitWithPosition 99.1->100.

+++ src/main/objlib.c
@@ -861,7 +861,7 @@ int ObjHits_GetPriorityHitWithPosition(int obj, int* outHitObject, int* outSpher
         for (hitSlot = 0; hitSlot < hitCount; hitSlot++)
         {
             hitPriority = hitState->priorities[hitSlot];
-            if ((s8)hitPriority < bestPriority)
+            if ((s8)hitPriority < (s8)bestPriority)
             {
                 bestPriority = hitPriority;
                 bestHitSlot = hitSlot;
@@ -917,7 +917,7 @@ int ObjHits_GetPriorityHit(int obj, int* outHitObject, int* outSphereIndex, u32*
         for (hitSlot = 0; hitSlot < hitCount; hitSlot++)
         {
             hitPriority = hitState->priorities[hitSlot];
-            if ((s8)hitPriority < bestPriority)
+            if ((s8)hitPriority < (s8)bestPriority)
             {
                 bestPriority = hitPriority;
                 bestHitSlot = hitSlot;
